Recuitment Type Validation done

+++ src/main/java/com/invicta/lms/repository/ProfileRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.invicta.lms.entity.Profile;
 
-public interface ProfileRepository extends JpaRepository<Profile, Integer> {
-	Profile findProfileById(Integer id);
+public interface ProfileRepository extends JpaRepository<Profile, Long> {
+	Profile findProfileById(Long id);
 
 }
